Separate client code and connect to tracker
